\name{bcont32}
\alias{bcont32}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized or unpenalized maximum likelihood optimization, when continuous margins (with three and two parameters, respectively) are employed.

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




