\name{bprob}
\alias{bprob}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Fisher information matrix for penalized or unpenalized maximum likelihood optimization.}



\author{

Giampiero Marra, University College London

Rosalba Radice, London School of Hygiene and Tropical Medicine

}



\references{

Marra G. and Radice R. (in press), Estimation of a Semiparametric Recursive Bivariate Probit in the Presence of Endogeneity. \emph{The Canadian 
Journal of Statistics}.

}









