\name{bprobgHsContUniv3}
\alias{bprobgHsContUniv3}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized or unpenalized maximum likelihood optimization, when fitting univariate models with continuous response based on three parameters.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






