\name{bprobgHsPL}
\alias{bprobgHsPL}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized or unpenalized maximum likelihood optimization, when using asymmetric link functions. Possible bivariate 
distributions are bivariate normal, Clayton, rotated Clayton (90 degrees), survival Clayton, rotated Clayton (270 degrees), 
Joe, rotated Joe (90 degrees), survival Joe, rotated Joe (270 degrees), Gumbel, rotated Gumbel (90 degrees), survival Gumbel, 
rotated Gumbel (270 degrees), and Frank. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\references{

Marra G. and Radice R. (2015), Flexible Bivariate Binary Models for Estimating the Efficacy of Phototherapy for Newborns with Jaundice. \emph{International Journal of Statistics and Probability}. 


}








