\name{bprobgHsPO}
\alias{bprobgHsPO}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed or expected information matrix for 
penalized or unpenalized maximum likelihood optimization for the case of bivariate normal models with partial
observability.  

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






