\name{bprobgHsPO0}
\alias{bprobgHsPO0}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed or expected information matrix for 
penalized or unpenalized maximum likelihood optimization for the case of bivariate models with partial
observability and zero correlation coefficient. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






