\name{bprobgHsSSBB}
\alias{bprobgHsSSBB}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized or unpenalized maximum likelihood optimization, for the non-random sample selection model case. 
Possible bivariate distributions are 
four Archimedean copulas with two parameters, namely the Clayton-Gumbel, Joe-Gumbel, Joe-Clayton and Joe-Frank. We refer to them
as BB1, BB6, BB7 and BB8, respectively. Their rotated versions are available as well.
}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






