\name{bprobNP.H}
\alias{bprobNP.H}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized 
or unpenalized maximum likelihood estimation in the presense
of nonparametric random effects. This is used for inferential purposes only.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\references{

Marra G., Papageorgiou G. and Radice R. (in press), Estimation of a Semiparametric Recursive Bivariate Probit Model with Nonparametric Mixing. \emph{Australian & New Zealand Journal of Statistics}.

}













