\name{bprobNRE}
\alias{bprobNRE}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed information matrix for 
penalized or unpenalized maximum likelihood optimization in the presense
of bivariate normal random effects.}



\author{

Georgios Papageorgiou (Imperial College London)

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}













