\name{bprobSS}
\alias{bprobSS}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Fisher information matrix for penalized or unpenalized maximum likelihood optimization, in the presence of sample selection.}



\author{

Giampiero Marra, University College London

Rosalba Radice, London School of Hygiene and Tropical Medicine

}


\references{


Marra G. and Radice R., A Penalized Likelihood Estimation Approach to Semiparametric Sample Selection Binary Response Modelling. \emph{Submitted}.


}









