\name{bprobSSH}
\alias{bprobSSH}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and hessian matrix for unpenalized maximum likelihood optimization, in the presence of sample selection.}



\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}







