\name{conv.check}
\alias{conv.check}

\title{Some convergence diagnostics}

\description{It takes a fitted \code{SemiParBIVProbit} object produced by \code{SemiParBIVProbit()} and produces some 
diagnostic information about the fitting procedure.

}

\usage{

conv.check(x)


}


\arguments{ 


\item{x}{\code{SemiParBIVProbit} object produced by \code{SemiParBIVProbit()}.}


}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{SemiParBIVProbit}}
}


\keyword{semiparametric bivariate modelling} 
\keyword{semiparametric trivariate probit modelling} 
\keyword{diagnostics} 







