\name{copgHs}
\alias{copgHs}

\title{Internal Function}

\description{This function evaluates the first and second derivatives with respect to the margins and association 
parameter of the bivariate normal, Clayton, rotated Clayton (90 degrees), survival Clayton, rotated Clayton (270 degrees), Joe, 
rotated Joe (90 degrees), survival Joe, rotated Joe (270 degrees), Gumbel, rotated Gumbel (90 degrees), survival Gumbel, 
rotated Gumbel (270 degrees), and Frank.


}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












