\name{copgHsAT}
\alias{copgHsAT}

\title{Internal Function}

\description{It evaluates the derivatives needed for the calculation of the average treatment effect in the case of continuous response for 
several copula functions.

}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












