\name{copgHsBB}
\alias{copgHsBB}

\title{Internal Function}

\description{This function evaluates the first and second derivatives with respect to the margins and association parameters of a given 
bivariate Archimedean copulas with two parameters.


}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












