\name{copgHsCont}
\alias{copgHsCont}

\title{Internal Function}

\description{This function evaluates the second and third derivatives with respect to the margins, dispersion and association 
parameters when employing the bivariate normal, Clayton, rotated Clayton (90 degrees), survival Clayton, rotated Clayton (270 degrees), Joe, 
rotated Joe (90 degrees), survival Joe, rotated Joe (270 degrees), Gumbel, rotated Gumbel (90 degrees), survival Gumbel, 
rotated Gumbel (270 degrees), and Frank.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












