\name{distrHs}
\alias{distrHs}

\title{Internal Function}

\description{It evaluates some of the derivatives needed in the likelihood function for the binary and continuous margin's case.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












