\name{distrHsAT}
\alias{distrHsAT}

\title{Internal Function}

\description{It evaluates the derivatives needed for the calculation of the average treatment effect in the case of continuous response.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












