\name{EBNormal}
\alias{EBNormal}

\title{Internal Function}

\description{It uses rejection sampling to obtain samples from the distribution of the random effects given the data. This is used when a model with bivariate normal random effects is fitted.}


\author{

Georgios Papageorgiou (Imperial College London)

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}













