\name{eta.tr}
\alias{eta.tr}
\alias{esp.tr}
\alias{enu.tr}
\alias{teta.tr}

\title{Internal Function}

\description{This and other similar internal functions map certain key quantities into a feasible parameter space.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






