\name{H.triSS}
\alias{H.triSS}

\title{Internal Function}

\description{It calculates the Hessian for a trivariate probit model with double sample selection.

}



\author{

Author: Panagiota Filippou

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








