\name{NP.qu}
\alias{NP.qu}

\title{Internal Function}

\description{It calcultes useful quantities (i.e. corrected Hessian matrix, estimated random effects, etc.) from a fitted bivariate probit model with nonparametric random effects.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\references{


Marra G., Papageorgiou G. and Radice R. (in press), Estimation of a Semiparametric Recursive Bivariate Probit Model with Nonparametric Mixing. \emph{Australian & New Zealand Journal of Statistics}.


}






