\name{penPL}
\alias{penPL}

\title{Internal Function}

\description{It provides an overall penalty matrix in a format suitable for estimation conditional on smoothing parameters, when
asymmetric links are used.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






