\name{print.AT}
\alias{print.AT}

\title{Print an AT object}

\description{The print method for a \code{AT} object.}

\usage{


\method{print}{AT}(x,...)


}


\arguments{ 


\item{x}{A \code{AT} object produced by \code{AT()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.AT} prints the lower confidence interval limit, estimated AT and upper confidence interval limit. 

}



\author{
Giampiero Marra, University College London

Rosalba Radice, London School of Hygiene and Tropical Medicine

}




\seealso{
\code{\link{AT}}
}


\keyword{semiparametric bivariate probit modelling} 








