\name{print.est.prev}
\alias{print.est.prev}

\title{Print an est.prev object}

\description{The print method for a \code{est.prev} object.}

\usage{


\method{print}{est.prev}(x,...)


}


\arguments{ 


\item{x}{A \code{est.prev} object produced by \code{est.prev()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.est.prev} prints the lower confidence interval limit, estimated prevalence and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\seealso{
\code{\link{est.prev}}
}

\keyword{semiparametric bivariate probit modelling} 









