\name{print.SemiParBIVProbit}
\alias{print.SemiParBIVProbit}

\title{Print a SemiParBIVProbit object}

\description{The print method for a \code{SemiParBIVProbit} object.}

\usage{


\method{print}{SemiParBIVProbit}(x,...)


}


\arguments{ 


\item{x}{A \code{SemiParBIVProbit} object produced by \code{SemiParBIVProbit()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.SemiParBIVProbit} prints out the family, model equations, total number of observations, estimated correlation 
coefficient and total effective degrees of freedom for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}



\seealso{
\code{\link{SemiParBIVProbit}}
}


\keyword{semiparametric bivariate probit modelling} 








