\name{print.SemiParTRIVProbit}
\alias{print.SemiParTRIVProbit}

\title{Print a SemiParTRIVProbit object}

\description{The print method for a \code{SemiParTRIVProbit} object.}

\usage{


\method{print}{SemiParTRIVProbit}(x, ...)


}


\arguments{ 


\item{x}{\code{SemiParTRIVProbit} object produced by \code{SemiParTRIVProbit()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.SemiParTRIVProbit} prints out the family, model equations, total number of observations, estimated association 
coefficient and total effective degrees of freedom for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{SemiParTRIVProbit}}
}


\keyword{semiparametric trivariate probit modelling} 








