\name{print.summary.copulaReg}
\alias{print.summary.copulaReg}

\title{Print a summary.copulaReg object}

\description{The print method for a \code{summary.copulaReg} object.}

\usage{


\method{print}{summary.copulaReg}(x, digits = max(3, getOption("digits") - 3), 
           signif.stars = getOption("show.signif.stars"), ...)


}


\arguments{ 


\item{x}{\code{summary.copulaReg} object produced by \code{summary.copulaReg()}.}

\item{digits}{Number of digits printed in output.} 

\item{signif.stars}{By default significance stars are printed alongside output.} 

\item{...}{Other arguments.}

}


\details{ 

\code{print.summary.copulaReg} prints model term summaries. 

}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{summary.copulaReg}}
}


\keyword{semiparametric bivariate probit modelling} 








