\name{print.summary.SemiParBIVProbit}
\alias{print.summary.SemiParBIVProbit}

\title{Print a summary.SemiParBIVProbit object}

\description{The print method for a \code{summary.SemiParBIVProbit} object.}

\usage{


\method{print}{summary.SemiParBIVProbit}(x,digits = max(3, getOption("digits") - 3), 
           signif.stars = getOption("show.signif.stars"),...)


}


\arguments{ 


\item{x}{A \code{summary.SemiParBIVProbit} object produced by \code{summary.SemiParBIVProbit()}.}

\item{digits}{Number of digits printed in output.} 

\item{signif.stars}{By default significance stars are printed alongside output.} 

\item{...}{Other arguments.}

}


\details{ 

\code{print.summary.SemiParBIVProbit} prints model term summaries. 

}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{summary.SemiParBIVProbit}}
}


\keyword{semiparametric bivariate probit modelling} 








