\name{S.m}
\alias{S.m}

\title{Internal Function}

\description{It provides penalty matrices in a format suitable for the automatic smoothness selection procedure.}


\author{
Giampiero Marra, University College London

Rosalba Radice, London School of Hygiene and Tropical Medicine

}






