\name{S.mPL}
\alias{S.mPL}

\title{Internal Function}

\description{It provides penalty matrices in a format suitable for the automatic smoothness selection procedure when a power link approach is used.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






