\name{sem.checks}
\alias{sem.checks}

\title{Some convergence diagnostics}

\description{It takes a fitted \code{SemiParBIVProbit} object produced by \code{SemiParBIVProbit()} and produces some diagnostic information about the fitting procedure.

}

\usage{

sem.checks(x)


}


\arguments{ 


\item{x}{A \code{SemiParBIVProbit} object produced by \code{SemiParBIVProbit()}.}


}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{SemiParBIVProbit}}
}


\keyword{semiparametric bivariate probit modelling} 








