\name{SemiParBIVProbit.fit.post}
\alias{SemiParBIVProbit.fit.post}
\alias{copulaReg.fit.post}
\alias{copulaSampleSel.fit.post}
\alias{SemiParTRIV.fit.post}

\title{Internal Function}

\description{This and other similar internal functions calculate useful post estimation quantities.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






