\name{spS}
\alias{spS}

\title{Internal Function}

\description{It provides penalty matrices in a format suitable for penalized likelihood optimization.}




\author{
Giampiero Marra, University College London

Rosalba Radice, London School of Hygiene and Tropical Medicine

}





