\name{startSS}
\alias{startSS}

\title{Internal Function}

\description{It calcultes sensible starting values, based on a two step procedure, for bivariate probit 
models with sample selection.}




\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}



\references{


Marra G. and Radice R., A Penalized Likelihood Estimation Approach to Semiparametric Sample Selection Binary Response Modelling. \emph{Submitted}.


}






