\name{triprobgHsSS}
\alias{triprobgHsSS}

\title{Internal Function}

\description{It provides score and Hessian for a trivariate probit model with double sample selection.
}



\author{

Author: Panagiota Filippou

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








