\name{working.comp}
\alias{working.comp}

\title{Internal Function}

\description{It efficiently calculates the working model quantities needed to implement the automatic multiple smoothing parameter 
procedure by exploiting the band structure of the weight matrix.}

\references{

Marra G. and Radice R. (in press), Estimation of a Semiparametric Recursive Bivariate Probit in the Presence of Endogeneity. The Canadian 
Journal of Statistics.

}

\author{
Giampiero Marra \email{giampiero@stats.ucl.ac.uk}
}








