\name{working.comp3}
\alias{working.comp3}

\title{Internal Function}

\description{It efficiently calculates the working model quantities needed to implement the automatic multiple smoothing parameter estimation
procedure by exploiting the band structure of the matrices involved in the calculations. This function includes the
association parameter of the two equations in the construction of the working linear model.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\references{

Radice R., Marra G. and M. Wojtys (submitted), Copula Regression Spline Models for Binary Outcomes.

}



