\name{working.compNP}
\alias{working.compNP}

\title{Internal Function}

\description{It efficiently calculates the working model quantities in the presence of nonparametric random effects 
needed to implement the automatic multiple smoothing parameter 
procedure.}




\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}






