\name{bprob}
\alias{bprob}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Fisher information matrix for penalized or unpenalized maximum likelihood optimization.}



\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}



\references{

Marra G. and Radice R. (2011), Estimation of a Semiparametric Recursive Bivariate Probit in the Presence of Endogeneity. \emph{Canadian 
Journal of Statistics}, 39(2), 259-279.

}









