\name{bprob}
\alias{bprob}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Fisher information matrix for penalized or unpenalized maximum likelihood optimization.}


\references{

Marra G. and Radice R. (in press), Estimation of a Semiparametric Recursive Bivariate Probit in the Presence of Endogeneity. The Canadian 
Journal of Statistics.

}


\author{

Rosalba Radice

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}








