\name{bprobgHsSS}
\alias{bprobgHsSS}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and observed or expected information matrix for 
penalized or unpenalized maximum likelihood optimization, for the non-random sample selection model case. Possible 
bivariate distributions are bivariate normal, Clayton, rotated Clayton (90 degrees), survival Clayton, rotated 
Clayton (270 degrees), Joe, rotated Joe (90 degrees), survival Joe, rotated Joe (270 degrees), Gumbel, rotated Gumbel (90 degrees), 
survival Gumbel, rotated Gumbel (270 degrees), Frank, FGM, AMH.

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}







