\name{bprobNP.H}
\alias{bprobNP.H}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and hessian matrix for penalized or unpenalized maximum likelihood estimation in the presense
of nonparametric random effects. This is used for inferential purposes only.}


\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}











