\name{bprobNP}
\alias{bprobNP}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Fisher information matrix for 
penalized or unpenalized maximum likelihood optimization in the presense
of nonparametric random effects.}



\author{

Georgios Papageorgiou (Imperial College London)

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\references{

Marra G., Papageorgiou G. and Radice R. (in press), Estimation of a Semiparametric Recursive Bivariate Probit Model with Nonparametric Mixing. \emph{Australian & New Zealand Journal of Statistics}.
}











