\name{bprobSS}
\alias{bprobSS}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Fisher information matrix 
for penalized or unpenalized maximum likelihood optimization, in the presence of non-random sample selection.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}


\references{


Marra G. and Radice R. (submitted), A Penalized Likelihood Estimation Approach to Semiparametric Sample Selection Binary Response Modeling. \emph{Electronic Journal of Statistics}, invited revision.

}









