\name{bprobSSH}
\alias{bprobSSH}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for unpenalized maximum likelihood optimization, in the presence of sample selection.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}







