\name{conv.check}
\alias{conv.check}

\title{Some convergence diagnostics}

\description{It takes a fitted model object and produces some 
diagnostic information about the fitting procedure.

}

\usage{

conv.check(x)


}


\arguments{ 


\item{x}{\code{SemiParBIVProbit}/\code{copulaReg}/\code{copulaSampleSel}/\code{SemiParTRIV} object.}


}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{SemiParBIVProbit}}, \code{\link{SemiParBIVProbit}}, \code{\link{copulaReg}}, \code{\link{copulaSampleSel}}, \code{\link{SemiParTRIV}}
}


\keyword{flexible copula regression modelling} 
\keyword{diagnostics} 







