\name{copgHs}
\alias{copgHs}

\title{Internal Function}

\description{This function evaluates the first and second derivatives with respect to the margins and association 
parameter of several bivariate copulas.


}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}












