\name{eta.tr}
\alias{eta.tr}
\alias{esp.tr}
\alias{enu.tr}
\alias{teta.tr}
\alias{dof.tr}
\alias{form.eq12}
\alias{ass.dp}
\alias{overall.sv}
\alias{overall.svG}
\alias{gamls.upsv}
\alias{pp}
\alias{susutsn}
\alias{susu}
\alias{pscr}
\alias{pscr0}
\alias{startsn}
\alias{bcorrec}
\alias{Cop1Cop2}
\alias{Reg2Copost}


\title{Internal Function}

\description{This and other similar internal functions map certain key quantities into a feasible parameter space.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






