\name{S.mPL}
\alias{S.mPL}

\title{Internal Function}

\description{It provides penalty matrices in a format suitable for automatic smoothness estimation, when an asymmetric link approach is used.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






