\name{startSS}
\alias{startSS}

\title{Internal Function}

\description{It calcultes sensible starting values for bivariate probit 
models with non-random sample selection. This is based on an adaptation of the Heckman two-step procedure.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\references{


Marra G. and Radice R. (2013), A Penalized Likelihood Estimation Approach to Semiparametric Sample Selection Binary Response Modeling. \emph{Electronic Journal of Statistics}, 7, 1432-1455.


}






