\name{working.comp}
\alias{working.comp}

\title{Internal Function}

\description{It efficiently calculates the working model quantities needed to implement the automatic multiple smoothing parameter 
procedure by exploiting the band structure of the weight matrix.}




\author{
Giampiero Marra, University College London

Rosalba Radice, London School of Hygiene and Tropical Medicine

}



\references{

Marra G. and Radice R. (2011), Estimation of a Semiparametric Recursive Bivariate Probit in the Presence of Endogeneity. \emph{Canadian 
Journal of Statistics}, 39, 259-279.

Marra G. and Radice R., A Penalized Likelihood Estimation Approach to Semiparametric Sample Selection Binary Response Modelling. \emph{Submitted}.


}






